requirements.env.txt
chardet==2.3.0
Django==1.10
django-debug-toolbar==1.6
django-filter==1.0.1
djangorestframework==3.5.4
docopt==0.6.2
language-selector==0.1
Markdown==2.6.8
pipreqs==0.4.6
psycopg2==2.6.2
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
requests==2.9.1
six==1.10.0
sqlparse==0.2.2
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
yarg==0.1.9
